The province is Navarra not Nafarroa
http://es.wikipedia.org/wiki/Navarra
The province is Navarra in Spanish, not Nafarroa, Nafarroa is Basque. 
The Government of Navarra website is: http://www.navarra.es
This correction and sent in previous versions, please correct it.
If each province appears in the local language, do not understand why still Navarra appearing as Nafarroa, it is incomprehensible, even if I myself had sent them this correction.